engine: major fixes and improvements
* fix compilation for win64
* add stb vorbis sound decoder for ogg
* add real sound streaming

+++ PHEngine/PHEngine/src/Core/UtilityCore/PlatformDependentFunctions.cpp
@@ -19,7 +19,7 @@ namespace EngineUtility
 
 	const char *get_module_file_name(HMODULE module)
 	{
-		size_t size = 1;
+		DWORD size = 1;
 		char *buffer;
 		for (;;)
 		{
